pyproject.toml
[tool.poetry]
name = "homun"
version = "0.1.0"
description = "A Discord bot for Backup Potion!"
authors = ["Floozutter <floozutter@gmail.com>"]
license = "Unlicense"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
